Refactored FFI and texture modules
