+++ src/ffi/cpp/etc1_encoder/include/etc1_encoder.h
