+++ src/ffi/cpp/etc1_encoder/include/rg_etc1.h
