+++ src/ffi/cpp/etc1_encoder/lib/release/etc1_encoder.lib
