+++ src/ffi/cpp/etc1_encoder/lib/release/libetc1_encoder.a
