+++ src/ffi/cpp/etc1_encoder/src/etc1_encoder.cpp
