+++ src/ffi/cpp/etc1_encoder/src/rg_etc1.cpp
